Extract test_seams.rs from turn.rs (parent #680)
Move the Issue #664 `#[cfg(test)] pub(super)` test seams out of
`turn.rs` into a focused `#![cfg(test)]` sibling module so `turn.rs`
keeps shrinking toward responsibility-focused units (parent meta-issue
#680).

Six test seams extracted as `#![cfg(test)]` free functions taking
`&mut Agent` / `&Agent` (matching `actor_loop_flow` / `reply_retry` /
earlier vertical-slice precedent):

- `effective_tool_policy_pub_for_test` (over private
  `effective_tool_policy()`)
- `build_arbiter_candidates_pub_for_test` (over private
  `build_arbiter_candidates()`)
- `drive_policy_error_for_test` (CB2-003: exercise
  `execute_tool_call`'s rejection + artifact-completion-attempt
  branches without network / cancellation / approval side effects)
- `artifact_directed_policy_for_test` (build
  `artifact_directed_from_job` from currently-installed
  `ArtifactCompletionJob`)
- `last_attempt_bash_policy_violation_for_test`
- `artifact_completion_job_attempts_len_for_test`

The new module uses `#![cfg(test)]` at the file head and the
registration in `loop_run.rs` is `#[cfg(test)] mod test_seams;`, so
the production binary excludes the module entirely. Also removed the
`#[cfg(test)] use super::tool_policy::effective_tool_policy_error_for_call_with_scope`
import in `turn.rs` — its only caller was `drive_policy_error_for_test`
and that import now lives in `test_seams.rs`.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `loop_run.rs` (6 sites: 5 `#[cfg(test)] pub(crate)` projection
seams + 1 inline test seam), `turn_tests.rs` (2 sites), and
`progress_tests.rs` (1 site) to the free-fn form.

LOC delta:
- turn.rs: 656 → 504 (-152)
- new module: +147

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
