+++ src/agent/loop_run/turn.rs
@@ -15,8 +15,6 @@ use super::tool_history::{
     is_preferred_read_edit_target, latest_user_turn_slice, recent_truncated_tool_call_attempt,
 };
 use super::tool_policy::EffectiveToolPolicy;
-#[cfg(test)]
-use super::tool_policy::effective_tool_policy_error_for_call_with_scope;
 use super::verifier_orchestration::task_contract_no_verifier_note;
 use super::workspace_walk::workspace_appears_empty;
 use super::*;
@@ -221,156 +219,6 @@ impl Agent {
             &self.models.main,
             self.plan_model_override.as_deref(),
         )
-    }
-
-    /// Issue #664 test seam: `pub(super)` wrapper over the private
-    /// `effective_tool_policy()` so `loop_run.rs`'s `#[cfg(test)]`
-    /// `effective_tool_policy_for_test` seam can reach the production
-    /// path without widening the `loop_run` module surface (DR3-001 /
-    /// AD19). Returns the raw `EffectiveToolPolicy`; the seam in
-    /// `loop_run.rs` projects it to `(Vec<String>, String)` primitives
-    /// before crossing the `pub(crate)` boundary.
-    #[cfg(test)]
-    pub(super) fn effective_tool_policy_pub_for_test(&self) -> EffectiveToolPolicy {
-        super::effective_tool_policy_flow::effective_tool_policy(self)
-    }
-
-    /// Issue #664 test seam: `pub(super)` wrapper over the private
-    /// `build_arbiter_candidates()` so the `loop_run.rs` seam can read
-    /// the candidate list without widening visibility. The returned
-    /// `Vec<JobCandidate>` stays inside `loop_run`; `loop_run.rs::
-    /// build_arbiter_candidates_for_test` projects each element into a
-    /// 4-field primitive DTO before the `pub(crate)` boundary.
-    #[cfg(test)]
-    pub(super) fn build_arbiter_candidates_pub_for_test(
-        &self,
-    ) -> Vec<super::active_job_arbiter::JobCandidate> {
-        super::effective_tool_policy_flow::build_arbiter_candidates(self)
-    }
-
-    /// Issue #664 iteration-3 (CB2-003) test seam: drive the
-    /// `effective_tool_policy_error_for_call_with_scope` rejection +
-    /// `record_artifact_completion_bash_violation` chokepoint under a
-    /// caller-provided `EffectiveToolPolicy` and tool arguments. Mirrors
-    /// the exact branching inside `execute_tool_call` without any
-    /// network / cancellation / approval side effects.
-    ///
-    /// Returns `(error_string, bash_violation_recorded_count_delta)`.
-    /// The count delta is observed by comparing the active job's
-    /// `attempts().len()` before/after the call, which captures both
-    /// the artifact-directed Bash branch (CB2-003) and the SetupBootstrap
-    /// branch (CB-003 iteration-2). The error string is the same value
-    /// `execute_tool_call` would surface to the LLM (sans
-    /// `lifecycle::format_tool_error` cosmetic wrapping).
-    ///
-    /// Visibility: `pub(super)` only; `loop_run.rs` projects this
-    /// through a primitive-tuple test seam.
-    #[cfg(test)]
-    pub(super) fn drive_policy_error_for_test(
-        &mut self,
-        policy: &EffectiveToolPolicy,
-        name: &str,
-        arguments: &serde_json::Value,
-    ) -> (Option<String>, usize) {
-        let before = self
-            .artifact_completion_job
-            .as_ref()
-            .map(|j| j.attempts().len())
-            .unwrap_or(0);
-        let scope_for_policy = if self.missing_verifier_job.is_some() {
-            Some(self.current_workspace_scope())
-        } else {
-            None
-        };
-        let err = effective_tool_policy_error_for_call_with_scope(
-            policy,
-            name,
-            arguments,
-            &self.work_root,
-            scope_for_policy.as_ref(),
-        );
-        if let Some(err) = err.as_ref() {
-            // Mirror the recording branch in `execute_tool_call` so the
-            // test seam exercises the exact production wiring.
-            if err.contains("artifact-directed recovery rejected") {
-                if name == "Bash" {
-                    let command_arg = arguments
-                        .get("command")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("")
-                        .to_string();
-                    let _ = super::artifact_completion_record::record_artifact_completion_bash_violation(self, vec![command_arg]);
-                } else {
-                    let actual_path = arguments
-                        .get("path")
-                        .and_then(serde_json::Value::as_str)
-                        .unwrap_or("")
-                        .to_string();
-                    let _ = super::artifact_completion_record::record_artifact_completion_attempt(
-                        self,
-                        super::artifact_completion_job::ArtifactAttemptOutcomeKind::WrongTarget,
-                        vec![format!("{name} on {actual_path}")],
-                    );
-                }
-            } else if err.starts_with("setup bootstrap")
-                && name == "Bash"
-                && self.artifact_completion_job.is_some()
-            {
-                let command_arg = arguments
-                    .get("command")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or("")
-                    .to_string();
-                let _ =
-                    super::artifact_completion_record::record_artifact_completion_bash_violation(
-                        self,
-                        vec![command_arg],
-                    );
-            }
-        }
-        let after = self
-            .artifact_completion_job
-            .as_ref()
-            .map(|j| j.attempts().len())
-            .unwrap_or(0);
-        (err, after.saturating_sub(before))
-    }
-
-    /// Issue #664 iteration-3 (CB2-003) test seam: build an
-    /// `artifact_directed_from_job` policy from the active
-    /// `ArtifactCompletionJob`. Returns `None` when no job is installed.
-    #[cfg(test)]
-    pub(super) fn artifact_directed_policy_for_test(&self) -> Option<EffectiveToolPolicy> {
-        let job = self.artifact_completion_job.as_ref()?;
-        let target = super::artifact_recovery_flow::artifact_recovery_target_path(self)?;
-        let target_already_read =
-            focused_edit_target_already_read(&self.session.messages, &target, &self.work_root);
-        Some(EffectiveToolPolicy::artifact_directed_from_job(
-            target,
-            target_already_read,
-            job.allowed_write_actions(),
-            job.allowed_read_scope(),
-        ))
-    }
-
-    /// Issue #664 iteration-3 (CB2-003) test seam: inspect the latest
-    /// recorded attempt's `bash_policy_violation` marker. `None` when no
-    /// job / no attempts. Mirrors the projection-time `category` field
-    /// without exposing `ArtifactAttemptOutcome`.
-    #[cfg(test)]
-    pub(super) fn last_attempt_bash_policy_violation_for_test(&self) -> Option<bool> {
-        let job = self.artifact_completion_job.as_ref()?;
-        let attempts = job.attempts();
-        attempts.last().map(|a| a.bash_policy_violation())
-    }
-
-    /// Issue #664 iteration-3 (CB2-003) test seam: count the number of
-    /// recorded attempts on the active job. `None` when no job.
-    #[cfg(test)]
-    pub(super) fn artifact_completion_job_attempts_len_for_test(&self) -> Option<usize> {
-        self.artifact_completion_job
-            .as_ref()
-            .map(|j| j.attempts().len())
     }
 
     pub(super) fn tool_specs_for_policy(&self, policy: &EffectiveToolPolicy) -> Vec<ToolSpec> {
